requirements.txt
flask==2.3.3
numpy==1.24.3
pillow==10.0.0
tensorflow==2.13.0
keras==2.13.1
requests==2.31.0
gunicorn==21.2.0
matplotlib==3.7.2
scikit-learn==1.3.0
tqdm==4.66.1
h5py==3.9.0